Push to update

+++ CnD_v1/CnD_v1/Entity/Enemy/LesserSkeleton.swift
@@ -73,7 +73,7 @@ class LesserSkeleton : Enemy
                         self.runAnimation(animAction: enemyAtkL!, animKey: "atkLeft")
                         if arc4random_uniform( (UInt32(50))) > player!.calcTotalEVS() / prs
                         {
-                            player!.dmgTaken += dmg / player!.calcTotalDEF()
+                            player!.dmgTaken += Float(dmg / player!.calcTotalDEF())
                         }
                     }
                 }
@@ -89,7 +89,7 @@ class LesserSkeleton : Enemy
                         self.runAnimation(animAction: enemyAtkR!, animKey: "atkRight")
                         if arc4random_uniform( (UInt32(50))) > player!.calcTotalEVS() / prs
                         {
-                            player!.dmgTaken += dmg / player!.calcTotalDEF()
+                            player!.dmgTaken += Float(dmg / player!.calcTotalDEF())
                         }
                     }
                 }
